+++ client/Packages/com.beamable/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Added
+
+- Add gift field in `Announcement` class.
+
 ## [1.16.0]
 
 ### Added